+++ tests/nested/subdir/B/main.cpp
@@ -15,7 +15,7 @@ int main(int argc, char* argv[]) {
   bar();
   four_bar();
   foo5(0);
-  foo6(0);
+  if (argc != 0) { if (argv[0] != 0) { foo6(0); } }
 
   return 0;
 }