+++ CHANGELOG.rst
@@ -30,7 +30,7 @@ Bug fixes and small improvements:
 
 - Fix warning ``Deprecated config key None used, please use 'txt-metric=branch' instead.``
   if ``txt-metric="branch"`` is used in config file. (:issue:`1066`)
-- Add ``excluded`` property for conditions and calls to the JSON report. (:issue:`1078`)
+- Add ``excluded`` property for conditions and calls to the JSON report. (:issue:`1080`)
 
 Documentation: